added templates, user profile

+++ base/views.py
@@ -1,11 +1,14 @@
 from django.contrib.auth import authenticate,login,logout
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.models import User
 from django.shortcuts import render, redirect
+from django.contrib.auth.decorators import login_required
 from django.http import HttpResponse
 from django.db.models import Q
-from .models import Room, Topic
+from .models import Room, Topic, Message
 from .forms import RoomForm
 from django.contrib import messages
+
 # Create your views here.
 
 #rooms = [
@@ -15,9 +18,14 @@ from django.contrib import messages
 #]
 
 def loginPage(request):
+    page = 'login'
     user = None  # Initialize the user variable
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
         try:
             user = User.objects.get(username=username)
@@ -31,15 +39,29 @@ def loginPage(request):
         else:
             messages.error(request, "Username or password is incorrect")
 
-    context = {}
+    context = {'page':page}
     return render(request, 'base/login_register.html', context)
-
-    context={}
-    return render(request,'base/login_register.html',context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    context = {'form':form}
+
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,"An error occured during registration.")
+    return render(request, 'base/login_register.html', context)
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -51,17 +73,37 @@ def home(request):
 
     topics = Topic.objects.all()
     room_count = rooms.count()
-    context = {'rooms':rooms,'topics':topics, 'room_count':room_count}
+    room_messages = Message.objects.filter(Q(room__topic__name__icontains=q))
+    context = {'rooms':rooms,'topics':topics, 'room_count':room_count, 'room_messages':room_messages}
     return render(request,'base/home.html',context)
 
 def room(request,pk):
     room = Room.objects.get(id=pk)
+    room_messages = room.message_set.all()
+    participants = room.participants.all()
 #    for i in rooms:
 #        if i['id'] == int(pk):
 #           room = i
-    context = {'room':room}
+    if request.method == "POST":
+        message = Message.objects.create(
+            user = request.user,
+            room = room,
+            body = request.POST.get('body')
+        )
+        room.participants.add(request.user)
+        return redirect ('room',pk=room.id)
+    context = {'room':room,'room_messages':room_messages,'participants':participants}
     return render(request,'base/room.html',context)
 
+def userProfile(request,pk):
+    user = User.objects.get(id=pk)
+    rooms = user.room_set.all()
+    room_messages = user.message_set.all()
+    topics = Topic.objects.all()
+    context={'user':user,'rooms':rooms,'topics':topics,'room_messages':room_messages}
+    return render(request,'base/profile.html',context)
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -72,9 +114,13 @@ def createRoom(request):
     context = {'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room) #pre-fill the existing data
+
+    if request.user != room.host: #cant update others' room
+        return HttpResponse('Get out!')
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance = room)
@@ -84,10 +130,27 @@ def updateRoom(request,pk):
     context = {'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Get out!')
+
     if request.method == 'POST': #post means we clicked confirm
         room.delete()
         return redirect("home")
     return render(request,'base/delete.html',{'obj':room})
+
+@login_required(login_url='login')
+def deleteMessage(request,pk):
+    message = Message.objects.get(id=pk)
+
+    if request.user != message.user:
+        return HttpResponse('Get out!')
+
+    if request.method == 'POST': #post means we clicked confirm
+        message.delete()
+        return redirect("home")
+    return render(request,'base/delete.html',{'obj':message})
